+++ libs/engine/CHANGELOG.rst
@@ -1,7 +1,7 @@
 **5.11.0 - 09/23/26**
 
 - Rename ``Pipeline.mutators`` to ``Pipeline.modifiers``
-- Add deprecation warning to ``Pipeline.mutators`` property
+- Add a ``FutureWarning`` to the deprecated ``Pipeline.mutators`` property
 - Assigning to ``Pipeline.mutators`` now raises ``AttributeError``; it was
   previously a plain writable attribute
 - Rename the ``mutator`` argument of the value combiners to ``modifier``